move function to get backdrop image to api-client utils
