+++ src/utils/url.ts
@@ -1,7 +1,3 @@
-import { BaseItemDto } from '@jellyfin/sdk/lib/generated-client';
-import { randomInt } from './number';
-import { ApiClient } from 'jellyfin-apiclient';
-
 /**
  * Gets the url search string.
  * This function should be used instead of location.search alone, because the app router
@@ -36,56 +32,4 @@ export const getParameterByName = (name: string, url?: string | null | undefined
 
     // eslint-disable-next-line compat/compat
     return new URLSearchParams(url).get(name) || '';
-};
-
-export interface ScaleImageOptions {
-    maxWidth?: number;
-    width?: number;
-    maxHeight?: number;
-    height?: number;
-    fillWidth?: number;
-    fillHeight?: number;
-    quality?: number;
-}
-
-/**
- * Returns the url of the first or a random backdrop image of an item.
- * If the item has no backdrop image, the url of the first or a random backdrop image of the parent item is returned.
- * Falls back to the primary image (cover) of the item, if neither the item nor it's parent have at least one backdrop image.
- * Returns undefined if no usable image was found.
- * @param apiClient The ApiClient to generate the url.
- * @param item The item for which the backdrop image is requested.
- * @param random If set to true and the item has more than one backdrop, a random image is returned.
- * @param options Optional; allows to scale the backdrop image.
- * @returns The url of the first or a random backdrop image of the provided item.
- */
-export const getItemBackdropImageUrl = (apiClient: ApiClient, item: BaseItemDto, random = false, options: ScaleImageOptions = {}): string | undefined => {
-    let imgUrl;
-
-    if (item.BackdropImageTags?.length) {
-        const backdropImgIndex = random ? randomInt(0, item.BackdropImageTags.length - 1) : 0;
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        imgUrl = apiClient.getScaledImageUrl(item.Id!, {
-            type: 'Backdrop',
-            index: backdropImgIndex,
-            tag: item.BackdropImageTags[backdropImgIndex],
-            ...options
-        });
-    } else if (item.ParentBackdropItemId && item.ParentBackdropImageTags?.length) {
-        const backdropImgIndex = random ? randomInt(0, item.ParentBackdropImageTags.length - 1) : 0;
-        imgUrl = apiClient.getScaledImageUrl(item.ParentBackdropItemId, {
-            type: 'Backdrop',
-            index: backdropImgIndex,
-            tag: item.ParentBackdropImageTags[backdropImgIndex],
-            ...options
-        });
-    } else if (item.ImageTags?.Primary) {
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        imgUrl = apiClient.getScaledImageUrl(item.Id!, {
-            type: 'Primary',
-            tag: item.ImageTags.Primary,
-            ...options
-        });
-    }
-    return imgUrl;
 };
